factored out user interest knn search
