+++ web-api/src/personalization.rs
@@ -11,6 +11,7 @@
 //
 // You should have received a copy of the GNU Affero General Public License
 
+mod personalized_knn;
 pub(crate) mod routes;
 
 use actix_web::web::ServiceConfig;
