+++ web-api/src/personalization/routes.rs
@@ -12,7 +12,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-use std::{cmp::Ordering, collections::HashMap, time::Duration};
+use std::{cmp::Ordering, collections::HashMap};
 
 use actix_web::{
     http::StatusCode,
@@ -22,17 +22,16 @@ use actix_web::{
     Responder,
 };
 use chrono::{DateTime, Utc};
-use futures_util::{stream::FuturesUnordered, StreamExt};
 use itertools::Itertools;
 use serde::{Deserialize, Serialize};
 use tracing::error;
-use xayn_ai_coi::{compute_coi_relevances, nan_safe_f32_cmp, CoiSystem, PositiveCoi};
+use xayn_ai_coi::{nan_safe_f32_cmp, CoiSystem};
 
-use super::{AppState, PersonalizationConfig, SemanticSearchConfig};
+use super::{personalized_knn, AppState, PersonalizationConfig, SemanticSearchConfig};
 use crate::{
     error::{
         application::WithRequestIdExt,
-        common::{BadRequest, DocumentNotFound, InternalError},
+        common::{BadRequest, DocumentNotFound},
     },
     models::{DocumentId, DocumentProperties, PersonalizedDocument, UserId, UserInteractionType},
     storage::{self, KnnSearchParams},
@@ -230,124 +229,6 @@ pub(crate) enum PersonalizedDocumentsError {
     NotEnoughInteractions,
 }
 
-/// Computes [`PositiveCoi`]s weights used to determine how many documents to fetch using each center's embedding.
-fn compute_coi_weights(cois: &[PositiveCoi], horizon: Duration, time: DateTime<Utc>) -> Vec<f32> {
-    let relevances = compute_coi_relevances(cois, horizon, time)
-        .into_iter()
-        .map(|rel| 1.0 - (-3.0 * rel).exp())
-        .collect_vec();
-
-    let relevance_sum = relevances.iter().sum::<f32>();
-    relevances
-        .iter()
-        .map(|relevance| {
-            let weight = relevance / relevance_sum;
-            if weight.is_finite() {
-                weight
-            } else {
-                // should be ok for our use-case
-                #[allow(clippy::cast_precision_loss)]
-                let len = cois.len() as f32;
-                // len can't be zero, because this iterator isn't entered for empty cois
-                1. / len
-            }
-        })
-        .collect()
-}
-
-/// Performs an approximate knn search for documents similar to the positive user interests.
-#[allow(clippy::too_many_arguments)]
-async fn search_knn_documents(
-    storage: &(impl storage::Document + storage::Interaction),
-    user_id: &UserId,
-    cois: &[PositiveCoi],
-    horizon: Duration,
-    time: DateTime<Utc>,
-    max_cois: usize,
-    store_user_history: bool,
-    count: usize,
-    published_after: Option<DateTime<Utc>>,
-) -> Result<Vec<PersonalizedDocument>, Error> {
-    let coi_weights = compute_coi_weights(cois, horizon, time);
-    let cois = cois
-        .iter()
-        .zip(coi_weights)
-        .sorted_by(|(_, a_weight), (_, b_weight)| nan_safe_f32_cmp(b_weight, a_weight))
-        .collect_vec();
-
-    let cois = &cois[..max_cois.min(cois.len())];
-    let weights_sum = cois.iter().map(|(_, w)| w).sum::<f32>();
-
-    let excluded = if store_user_history {
-        storage::Interaction::get(storage, user_id).await?
-    } else {
-        Vec::new()
-    };
-
-    let mut document_futures = cois
-        .iter()
-        .map(|(coi, weight)| async {
-            // weights_sum can't be zero, because coi weights will always return some weights that are > 0
-            let weight = *weight / weights_sum;
-            #[allow(
-            // fine as max documents count is small enough
-            clippy::cast_precision_loss,
-            // fine as weight should be between 0 and 1
-            clippy::cast_sign_loss,
-            // fine as number of neighbors is small enough
-            clippy::cast_possible_truncation
-        )]
-            let k_neighbors = (weight * count as f32).ceil() as usize;
-
-            storage::Document::get_by_embedding(
-                storage,
-                KnnSearchParams {
-                    excluded: &excluded,
-                    embedding: &coi.point,
-                    k_neighbors,
-                    num_candidates: count,
-                    published_after,
-                    min_similarity: None,
-                    time,
-                },
-            )
-            .await
-        })
-        .collect::<FuturesUnordered<_>>();
-
-    let mut all_documents = HashMap::new();
-    let mut errors = Vec::new();
-
-    while let Some(result) = document_futures.next().await {
-        match result {
-            Ok(documents) => {
-                // the same document can be returned with different elastic scores, hence the
-                // documents are deduplicated and only the highest score is retained for each
-                for document in documents {
-                    all_documents
-                        .entry(document.id.clone())
-                        .and_modify(|PersonalizedDocument { score, .. }| {
-                            if *score < document.score {
-                                *score = document.score;
-                            }
-                        })
-                        .or_insert(document);
-                }
-            }
-            Err(error) => {
-                error!("Error fetching documents: {error}");
-                errors.push(error);
-            }
-        };
-    }
-
-    if all_documents.is_empty() && !errors.is_empty() {
-        Err(InternalError::from_message("Fetching documents failed").into())
-    } else {
-        Ok(all_documents.into_values().collect_vec())
-    }
-}
-
 pub(crate) enum PersonalizeBy<'a> {
     KnnSearch {
         count: usize,
@@ -372,22 +253,26 @@ pub(crate) async fn personalize_documents_by(
         return Ok(None);
     }
 
+    let excluded = if personalization.store_user_history {
+        storage::Interaction::get(storage, user_id).await?
+    } else {
+        Vec::new()
+    };
+
     let mut all_documents = match by {
         PersonalizeBy::KnnSearch {
             count,
             published_after,
         } => {
-            search_knn_documents(
-                storage,
-                user_id,
-                &cois.positive,
-                coi.config().horizon(),
-                time,
-                personalization.max_cois_for_knn,
-                personalization.store_user_history,
+            personalized_knn::Search {
+                interests: &cois.positive,
+                excluded: &excluded,
+                horizon: coi.config().horizon(),
+                max_cois: personalization.max_cois_for_knn,
                 count,
                 published_after,
-            )
+            }
+            .run_on(storage)
             .await?
         }
         PersonalizeBy::Documents(documents) => {
@@ -507,34 +392,4 @@ async fn semantic_search(
     Ok(Json(SemanticSearchResponse {
         documents: documents.into_iter().map_into().collect(),
     }))
-}
-
-#[cfg(test)]
-mod tests {
-    use xayn_ai_coi::CoiConfig;
-
-    use super::*;
-    use crate::storage::memory::Storage;
-
-    #[tokio::test]
-    async fn test_search_knn_documents_for_empty_cois() {
-        // these dummy values are only used for the excluded documents which will be empty
-        let storage = Storage::default();
-        let user = "123".try_into().unwrap();
-
-        let documents = search_knn_documents(
-            &storage,
-            &user,
-            &[],
-            CoiConfig::default().horizon(),
-            Utc::now(),
-            PersonalizationConfig::default().max_cois_for_knn,
-            PersonalizationConfig::default().store_user_history,
-            10,
-            None,
-        )
-        .await
-        .unwrap();
-        assert!(documents.is_empty());
-    }
 }
